render the empty state template parts for the Mini Cart Block #4674
render the empty state template parts for the Mini Cart Block #4674
