--- assets/js/blocks/cart-checkout/mini-cart-contents/inner-blocks/filled-mini-cart-contents-block/block.tsx
@@ -1,5 +1,0 @@
-const Block = ( { children }: { children: JSX.Element } ): JSX.Element => {
-	return <>{ children }</>;
-};
-
-export default Block;
