+++ packages/checkout/blocks-registry/types.ts
@@ -18,6 +18,7 @@ export enum innerBlockAreas {
 	FILLED_CART = 'woocommerce/filled-cart-block',
 	CART_ITEMS = 'woocommerce/cart-items-block',
 	CART_TOTALS = 'woocommerce/cart-totals-block',
+	MINI_CART_CONTENTS = 'woocommerce/mini-cart-contents',
 	EMPTY_MINI_CART = 'woocommerce/empty-mini-cart-contents-block',
 	FILLED_MINI_CART = 'woocommerce/filled-mini-cart-content-block',
 }
